Fix : Playwright Flaky Tests (#1729)
* Increase Workers

* Add debug statements

* Updated metrics Test

* All Alert Update

* Metrics Alert Fix

* Divided Slow Test in Two Files

* Increase Timeout

* Updated All Alert Test

* Alert Test Files

* Yaml Change

* Dashboards Home Test Update

* Update workers

* Update workers

* Debugging test

* grid container

* Fix

* increase timeout

* Cleanup

* Increase Timeout

* Fix

* Setting `CI: true`
